pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = ""
authors = ["anup.gogoi <anup.gogoi@berrijam.com>"]

[tool.poetry.dependencies]
python = "^3.9.0"

django = "*"
djangorestframework = "*"
djangorestframework_simplejwt = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
